Refactor the unit ball examples
