+++ drake/solvers/test/convex_optimization_test.cc
@@ -21,13 +21,6 @@ namespace drake {
 namespace solvers {
 namespace test {
 namespace {
-void GetQuadraticProgramSolvers(
-    std::list<std::unique_ptr<MathematicalProgramSolverInterface>>* solvers) {
-  AddSolverIfAvailable(SolverType::kGurobi, solvers);
-  AddSolverIfAvailable(SolverType::kMosek, solvers);
-  AddSolverIfAvailable(SolverType::kSnopt, solvers);
-}
-
 void GetSecondOrderConicProgramSolvers(
     std::list<std::unique_ptr<MathematicalProgramSolverInterface>>* solvers) {
   AddSolverIfAvailable(SolverType::kGurobi, solvers);
@@ -37,112 +30,6 @@ void GetSecondOrderConicProgramSolvers(
 void GetSemidefiniteProgramSolvers(
     std::list<std::unique_ptr<MathematicalProgramSolverInterface>>* solvers) {
   AddSolverIfAvailable(SolverType::kMosek, solvers);
-}
-
-// Test the simple QP
-// min x(0)^2 + x(1)^2 + 2 * x(2)^2
-// s.t x(0) +   x(1) = 1
-//     x(0) + 2*x(2) = 2
-// The optimal solution should be
-// x(0) = 4/5, x(1) = 1/5, x(2) = 3/5
-void TestQuadraticProgram4(const MathematicalProgramSolverInterface& solver) {
-  MathematicalProgram prog;
-  auto x = prog.NewContinuousVariables(3);
-
-  Eigen::Matrix3d Q = Eigen::Matrix3d::Identity();
-  Q(2, 2) = 2.0;
-  Eigen::Vector3d b = Eigen::Vector3d::Zero();
-  prog.AddQuadraticCost(Q, b, x);
-  prog.AddLinearEqualityConstraint(Eigen::RowVector2d(1, 1),
-                                   Vector1d::Constant(1), x.head<2>());
-  prog.AddLinearEqualityConstraint(Eigen::RowVector2d(1, 2),
-                                   Vector1d::Constant(2),
-                                   {x.segment<1>(0), x.segment<1>(2)});
-
-  if (solver.solver_type() == SolverType::kSnopt) {
-    prog.SetInitialGuessForAllVariables(Eigen::Vector3d::Zero());
-  }
-  RunSolver(&prog, solver);
-  const auto& x_value = prog.GetSolution(x);
-  EXPECT_TRUE(CompareMatrices(Eigen::Vector3d(0.8, 0.2, 0.6), x_value, 1e-9,
-                              MatrixCompareType::absolute));
-}
-
-// Solve a series of QPs with the objective being the Euclidean distance
-// from a point which moves along the unit circle (L2 ball), but
-// constrained to lie inside the L1 ball.  Implemented in 2D, so that the
-// active set moves along 4 faces of the L1 ball.
-void TestQPonUnitBallExample(const MathematicalProgramSolverInterface& solver) {
-  MathematicalProgram prog;
-  auto x = prog.NewContinuousVariables(2);
-
-  Eigen::Matrix2d Q = Eigen::Matrix2d::Identity();
-  Eigen::Vector2d x_desired;
-  x_desired << 1.0, 0.0;
-  auto objective = prog.AddQuadraticErrorCost(Q, x_desired, x);
-
-  Eigen::Matrix2d A;
-  A << 1.0, 1.0, -1.0, 1.0;
-  Eigen::Vector2d ub = Eigen::Vector2d::Constant(1.0);
-  Eigen::Vector2d lb = Eigen::Vector2d::Constant(-1.0);
-  auto constraint = prog.AddLinearConstraint(A, lb, ub, x);
-  Eigen::Vector2d x_expected;
-
-  const int N = 40;  // number of points to test around the circle
-  for (int i = 0; i < N; i++) {
-    double theta = 2.0 * M_PI * i / N;
-    x_desired << sin(theta), cos(theta);
-    objective->UpdateQuadraticAndLinearTerms(2.0 * Q, -2.0 * Q * x_desired);
-
-    if (theta <= M_PI_2) {
-      // simple lagrange multiplier problem:
-      // min (x-x_d)^2 + (y-y_d)^2 s.t. x+y=1
-      x_expected << (x_desired(0) - x_desired(1) + 1.0) / 2.0,
-          (x_desired(1) - x_desired(0) + 1.0) / 2.0;
-    } else if (theta <= M_PI) {
-      // min (x-x_d)^2 + (y-y_d)^2 s.t. x-y=1
-      x_expected << (x_desired(0) + x_desired(1) + 1.0) / 2.0,
-          (x_desired(0) + x_desired(1) - 1.0) / 2.0;
-    } else if (theta <= 3.0 * M_PI_2) {
-      // min (x-x_d)^2 + (y-y_d)^2 s.t. x+y=-1
-      x_expected << (x_desired(0) - x_desired(1) - 1.0) / 2.0,
-          (x_desired(1) - x_desired(0) - 1.0) / 2.0;
-    } else {
-      // min (x-x_d)^2 + (y-y_d)^2 s.t. x-y=-1
-      x_expected << (x_desired(0) + x_desired(1) - 1.0) / 2.0,
-          (x_desired(0) + x_desired(1) + 1.0) / 2.0;
-    }
-
-    if (solver.solver_type() == SolverType::kSnopt) {
-      prog.SetInitialGuessForAllVariables(Eigen::Vector2d::Zero());
-    }
-    RunSolver(&prog, solver);
-    const auto& x_value = prog.GetSolution(x);
-
-    EXPECT_TRUE(CompareMatrices(x_value, x_expected, 1e-4,
-                                MatrixCompareType::absolute));
-  }
-
-  // provide some test coverage for changing Q
-  //
-  {
-    // now 2(x-xd)^2 + (y-yd)^2 s.t. x+y=1
-    x_desired << 1.0, 1.0;
-    Q(0, 0) = 2.0;
-    objective->UpdateQuadraticAndLinearTerms(2.0 * Q, -2.0 * Q * x_desired);
-
-    x_expected << 2.0 / 3.0, 1.0 / 3.0;
-
-    SolutionResult result = SolutionResult::kUnknownError;
-
-    prog.SetSolverOption(SolverType::kGurobi, "BarConvTol", 1E-9);
-    ASSERT_NO_THROW(result = prog.Solve());
-    EXPECT_EQ(result, SolutionResult::kSolutionFound);
-
-    const auto& x_value = prog.GetSolution(x);
-    EXPECT_TRUE(CompareMatrices(x_value, x_expected, 1e-5,
-                                MatrixCompareType::absolute));
-  }
 }
 
 ////////////////////////////////////////
@@ -567,118 +454,6 @@ void TestFindSpringEquilibrium(
                         solver);
 }
 }  // namespace
-
-GTEST_TEST(TestQP, TestL2NormCost) {
-  MathematicalProgram prog;
-  auto x = prog.NewContinuousVariables<2>();
-
-  // |Ax - b|^2 = (x-xd)'Q(x-xd) => Q = A'*A and b = A*xd.
-  Eigen::Matrix2d A;
-  A << 1, 2, 3, 4;
-  Eigen::Matrix2d Q = A.transpose() * A;
-  Eigen::Vector2d x_desired;
-  x_desired << 5, 6;
-  Eigen::Vector2d b = A * x_desired;
-
-  std::shared_ptr<QuadraticConstraint> obj1 =
-      prog.AddQuadraticErrorCost(Q, x_desired, x);
-  std::shared_ptr<QuadraticConstraint> obj2 = prog.AddL2NormCost(A, b, x);
-
-  // Test the objective at a 6 arbitrary values (to guarantee correctness
-  // of the six-parameter quadratic form.
-  Eigen::Vector2d x0;
-  Eigen::VectorXd y1, y2;
-  x0 << 7, 8;
-
-  for (int i = 0; i < 6; i++) {
-    obj1->Eval(x0, y1);
-    obj2->Eval(x0, y2);
-
-    EXPECT_TRUE(CompareMatrices(y1, y2));
-    EXPECT_TRUE(CompareMatrices(
-        y2, (A * x0 - b).transpose() * (A * x0 - b) - b.transpose() * b));
-    // Note: Currently have to subtract out the constant term (b'*b) due to
-    // issue #3500.
-
-    x0 += Eigen::Vector2d::Constant(2);
-  }
-}
-
-GTEST_TEST(TestQP, TestQuadraticProgram4) {
-  std::list<std::unique_ptr<MathematicalProgramSolverInterface>> solvers;
-  GetQuadraticProgramSolvers(&solvers);
-  for (const auto& solver : solvers) {
-    TestQuadraticProgram4(*solver);
-  }
-}
-
-GTEST_TEST(TestQP, TestQuadraticProgramUnitBall) {
-  std::list<std::unique_ptr<MathematicalProgramSolverInterface>> solvers;
-  GetQuadraticProgramSolvers(&solvers);
-  for (const auto& solver : solvers) {
-    TestQPonUnitBallExample(*solver);
-  }
-}
-
-GTEST_TEST(TestQP, TestQuadraticProgram5) {
-  std::list<std::unique_ptr<MathematicalProgramSolverInterface>> solvers;
-  GetQuadraticProgramSolvers(&solvers);
-  Eigen::VectorXd solution;
-  bool get_solution = false;
-  for (const auto& solver : solvers) {
-    MathematicalProgram prog;
-    auto x = prog.NewContinuousVariables<2>("x");
-    auto z = prog.NewContinuousVariables<2>("z");
-    auto y = prog.NewContinuousVariables<2>("y");
-
-    // Randomly generated vector.
-    Eigen::Matrix<double, 2, 4> CE_xy;
-    CE_xy << 0.3147, 0.1324, 0.4575, 0.4572, 0.4058, -0.4025, 0.4649, -0.0146;
-    Eigen::Matrix<double, 2, 1> ce0_xy;
-    ce0_xy << -0.0782, 0.4157;
-
-    Eigen::Matrix<double, 1, 4> CI_xy;
-    CI_xy << 0.1797, 0.0853, 0.4593, 0.3143;
-    Eigen::Matrix<double, 1, 1> ci_xy_upper, ci_xy_lower;
-    ci_xy_upper << -10;
-    ci_xy_lower << -std::numeric_limits<double>::infinity();
-
-    Eigen::Matrix<double, 1, 2> CI_z;
-    CI_z << -0.3639, 0.3693;
-    Eigen::Matrix<double, 1, 1> ci_z_upper, ci_z_lower;
-    ci_z_upper << -10;
-    ci_z_lower << -std::numeric_limits<double>::infinity();
-
-    // Randomly generated matrix.
-    Eigen::Matrix<double, 6, 6> G;
-    G << 0.6472, -0.1517, -0.1019, -0.4394, 0.1772, 0.1697, -0.1517, 0.3755,
-        0.0967, 0.3431, 0.2979, 0.1124, -0.1019, 0.0967, 0.1902, 0.0987, 0.2064,
-        -0.0820, -0.4394, 0.3431, 0.0987, 0.6460, 0.1524, 0.0546, 0.1772,
-        0.2979, 0.2064, 0.1524, 1.0408, 0.2175, 0.1697, 0.1124, -0.0820, 0.0546,
-        0.2175, 0.5526;
-
-    Eigen::Matrix<double, 6, 1> g0;
-    g0 << 0.3001, -0.0686, 0.4106, -0.3182, -0.2362, -0.3545;
-
-    prog.AddQuadraticCost(G, g0, {x, z, y});
-    prog.AddLinearEqualityConstraint(CE_xy, ce0_xy, {x, y});
-    prog.AddLinearConstraint(CI_xy, ci_xy_lower, ci_xy_upper, {x, y});
-    prog.AddLinearConstraint(CI_z, ci_z_lower, ci_z_upper, z);
-
-    if (solver->solver_type() == SolverType::kSnopt) {
-      prog.SetInitialGuessForAllVariables(drake::Vector6<double>::Zero());
-    }
-    RunSolver(&prog, *solver);
-
-    if (get_solution) {
-      EXPECT_TRUE(CompareMatrices(solution, prog.GetSolutionVectorValues(),
-                                  1E-4, MatrixCompareType::absolute));
-    } else {
-      solution = prog.GetSolutionVectorValues();
-      get_solution = true;
-    }
-  }
-}
 
 GTEST_TEST(TestSOCP, TestEllipsoidsSeparation0) {
   std::list<std::unique_ptr<MathematicalProgramSolverInterface>> solvers;
