+++ drake/solvers/test/gurobi_solver_test.cc
@@ -61,6 +61,10 @@ INSTANTIATE_TEST_CASE_P(
                        ::testing::ValuesIn(linear_constraint_form()),
                        ::testing::ValuesIn(quadratic_problems())));
 
+GTEST_TEST(QPtest, TestUnitBallExample) {
+  GurobiSolver solver;
+  TestQPonUnitBallExample(solver);
+}
 }  // namespace test
 }  // namespace solvers
 }  // namespace drake
